separate routes and update db methods
